+++ grails-app/controllers/io/xh/hoist/impl/XhController.groovy
@@ -166,7 +166,6 @@ class XhController extends BaseController {
             prefService.clearPreferences()
             viewService.clearAllState()
         }
-
         renderSuccess()
     }